refactor: move default-value-config into default-value core module folder
