+++ src/app/core/entity/default-datatype/edit-component.ts
@@ -2,7 +2,7 @@ import { FormControl, FormGroup } from "@angular/forms";
 import { FormFieldConfig } from "../../common-components/entity-form/FormConfig";
 import { Entity } from "../model/entity";
 import { Directive, Input, OnChanges, OnInit } from "@angular/core";
-import { DefaultValueConfig } from "../schema/default-value-config";
+import { DefaultValueConfig } from "../../default-values/default-value-config";
 import { EntityForm } from "app/core/common-components/entity-form/entity-form.service";
 
 /**
